requirements.txt
dashdash==2.15
pandas==2.2.1
dash-bootstrap-components=1.5